requirements.txt
python-telegram-bot[job-queue]~=21.0
SQLAlchemy>=1.4
Pillow==10.0.0
imagehash==4.2.0
scipy>=1.1.0
python-dotenv==1.0.0
